Final fixes, EU finished

+++ s3_to_ceph.rb
@@ -20,11 +20,37 @@ require 'bundler/setup'
 require 'thread_safe'
 require 'logger'
 require 'yaml'
-require 'fog'
 require 'pry'
 require 'tempfile'
 require 'parallel'
 require 'timeout'
+require 'mime-types'
+
+module MIME
+  class Types
+    def initialize(data_version = nil)
+      @type_variants    = ThreadSafe::Hash.new { |h, k| h[k] = [] }
+      @extension_index  = ThreadSafe::Hash.new
+      @extension_index['jpg'] = ThreadSafe::Array.new 
+      @extension_index['jpeg'] = ThreadSafe::Array.new 
+      @extension_index['png'] = ThreadSafe::Array.new 
+      @extension_index['gif'] = ThreadSafe::Array.new 
+      @extension_index['tiff'] = ThreadSafe::Array.new 
+      @extension_index['tif'] = ThreadSafe::Array.new 
+      @data_version = data_version
+    end
+    
+    def type_for(filename, platform = false)
+      ext = filename.chomp.downcase.gsub(/.*\./o, '')
+      list = @extension_index.clone[ext]
+      list.delete_if { |e| not e.platform? } if platform
+      list
+    end
+  end
+end
+
+require 'fog'
+
 # No longer necesarry for next Fog release
 module Excon
   class Connection
@@ -36,18 +62,7 @@ Excon.defaults[:ssl_verify_peer] = false
 
 OpenSSL::SSL::VERIFY_PEER = OpenSSL::SSL::VERIFY_NONE
 
-module MIME
-  class Types
-    def initialize(data_version = nil)
-      @type_variants    = ThreadSafe::Hash.new { |h, k| h[k] = [] }
-      @extension_index  = ThreadSafe::Hash.new { |h, k| h[k] = [] }
-      @data_version = data_version
-    end
-  end
-end
-
 module S3Backup
-
   mylogger = Logger.new('backup.log', 'daily')
   STOR_CONF = {}
 
@@ -83,26 +98,26 @@ module S3Backup
   ceph_dir = ceph.directories.get('images.eu.viewbook.com')
   files = s3.directories.get('images.eu.viewbook.com').files
 
-
-  subset = files.all(:marker => '56c33849662fab07a58bb76a5f083bfa.jpg')
+  subset = files.all(:marker => 'fffffeacf7afed9b3de61fd3a01776f7.jpg')
 
   def S3Backup.parallel_copy files, target_dir, logger
     Parallel.each(files, :in_threads => 12) do |s3_file|
+    next if s3_file.key.include?('/') # can't handle slashes in filenames yet
     begin
     Timeout.timeout(60) do
       logger.info(s3_file.key)
       tempfile = Tempfile.new(s3_file.key)
-      try_this(3, "S3 error") do
+      try_this(3, "S3") do
         tempfile.write(s3_file.body)
       end
-      try_this(3, "Ceph error") do
+      try_this(3, "Ceph") do
         target_dir.files.create(:key => s3_file.key, :body => tempfile, :public => true )
       end
       tempfile.close
       tempfile.unlink
     end
     rescue Timeout::Error
-        $stderr.puts("Timeout")
+      $stderr.puts("Timeout")
     end
     end
   end
@@ -110,13 +125,14 @@ module S3Backup
   def S3Backup.try_this tries=3, error="error"
     begin
       yield
-    rescue
+    rescue Excon::Errors::InternalServerError => e
       tries += 1
       if tries < 3
-        sleep 3
+        sleep 30
         retry
       else
-        $stderr.puts(error)
+        $stderr.puts("#{error}: #{e.message}")
+	#raise "Too many errors"
       end
     end
   end
